Bioconductor packages using CRAN archived packages

Checks on the 4 Bioconductor repositories which packages depend on a archived package.

Usage

bioc_cran_archived(which = "strong")

Arguments

which

a character vector listing the types of dependencies, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"). Character string "all" is shorthand for that vector, character string "most" for the same vector without "Enhances", character string "strong" (default) for the first three elements of that vector.

Value

A data.frame with the name of the Bioconductor packages depending on archived packages (on Archived column) and the number of missing packages (n). NA if not able to collect the data.

See also

Examples

bca <- bioc_cran_archived()
#> Retrieving bioc_available_release, this might take a bit.
#> Caching results to be faster next call in this session.
#> Retrieving CRAN_db, this might take a bit.
#> Caching results to be faster next call in this session.
head(bca)
#>     Package  Archived n
#> 1     BiGGR       LIM 1
#> 2     ChAMP      kpmt 1
#> 3    ChIPQC Nozzle.R1 1
#> 4  PhenStat SmoothWin 1
#> 5     RITAN  linkcomm 1
#> 6 Repitools  gsmoothr 1