Retrieve CRAN archive

Retrieve the archive and the current database.

Usage

cran_archive(packages = NULL)

Arguments

packages

A vector with packages or NULL for all packages.

Value

A data.frame with 6 columns: Package, Date (of publication), Version, User, size and status (archived or current). It is sorted by package name and date.

Details

Some packages would get an NA in Version, if package_version() were to be used with strict = FALSE. Packages might have been on CRAN but could have been removed and won't show up. Depending on the data requested and packages currently on CRAN, you might get a warning regarding a package being both archived and current.

See also

Examples

# \donttest{
ap <- available.packages()
a_package <- rownames(ap)[startsWith(rownames(ap), "A")][2]
ca <- cran_archive(a_package)
head(ca)
#>    Package            Datetime Version   User   Size   Status
#> 1 AATtools 2020-06-14 15:10:07   0.0.1 ligges 145230 archived
#> 2 AATtools 2022-08-12 13:40:11   0.0.2 ligges 249141 archived
#> 3 AATtools 2024-08-16 09:10:08   0.0.3 ligges 252641  current
# }