CRAN comments

CRAN volunteers document since ~2009 why they archive packages. This function retrieves the data and prepares it for analysis, classifying the actions taken by the team per package and date.

Usage

cran_comments(packages = NULL)

Arguments

packages

A vector with packages or NULL for all packages.

Value

A data.frame with four columns: package, comment, date and action.

Details

The comments are slightly edited: multiple comments for the same action are joined together so that they can be displayed on a single line. Actions are inferred from 7 keywords: archived, orphaned, removed, renamed, replaced, unarchived, unorphaned.

Note

There can be room for improvement: some comments describe two actions, please let me know if you think this can be improved. Other actions can be described on multiple comments/lines or out of order. Compare with the original file in case of doubts.

References

Original file: https://cran.r-project.org/src/contrib/PACKAGES.in

Examples

# \donttest{
cc <- cran_comments()
#> Retrieving comments, this might take a bit. Next call will be faster.
head(cc)
#>       package
#> 1 ABCExtremes
#> 2       ABCp2
#> 3       ABCp2
#> 4      ACCLMA
#> 5         ACD
#> 6        ACDC
#>                                                                                            comment
#> 1     Archived on 2015-06-19 as incomplete maintainer address was not corrected despite reminders.
#> 2                     Archived on 2015-07-01 as maintainer address <duryea@dartmouth.edu> bounced.
#> 3                           Archived on 2025-01-26 as issues were not corrected despite reminders.
#> 4 Archived on 2021-02-05 as issues were not corrected despite reminders: 'xtfrm()' on data frames.
#> 5                   Archived on 2022-06-08 as check problems were not corrected despite reminders.
#> 6                                         Archived on 2023-05-09 at the request of the maintainer.
#>         date   action
#> 1 2015-06-19 archived
#> 2 2015-07-01 archived
#> 3 2025-01-26 archived
#> 4 2021-02-05 archived
#> 5 2022-06-08 archived
#> 6 2023-05-09 archived
# }