Calculate time till packages are archived

Given the deadlines by the CRAN volunteers packages can be archived which can trigger some other packages to be archived. This code calculates how much time the chain reaction will go on if maintainer don't fix/update the packages.

Usage

cran_doom(which = "strong", bioc = FALSE)

Arguments

which

a character vector listing the types of dependencies, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"). Character string "all" is shorthand for that vector, character string "most" for the same vector without "Enhances", character string "strong" (default) for the first three elements of that vector.

bioc

Logical value if Bioconductor packages should be provided, (Requires internet connection).

Value

A list with multiple elements:

time_till_last: Time till last package is affected.

last_archived: the date of the last package that would be affected.

npackages: Numeric vector with the number of packages used.

details: A data.frame with information for each individual package: Name, date affected, affected directly, repository, times it is affected (by archival causing issues.) NA if not able to collect the data from CRAN.

Details

Packages on Suggested: field should

References

Original code from: https://github.com/schochastics/cran-doomsday/blob/main/index.qmd

See also

Examples

# \donttest{
cd <- cran_doom()
if (length(cd) > 1L) head(cd$details)
#>        Package   Deadline   type repo n_affected
#> 1      kvkapiR 2025-09-22 direct CRAN          1
#> 2    metacoder 2025-09-25 direct CRAN          5
#> 3    DescrTab2 2025-09-25 direct CRAN          4
#> 4 OlinkAnalyze 2025-09-25 direct CRAN          4
#> 5         fgdr 2025-09-25 direct CRAN          4
#> 6        jstor 2025-09-25 direct CRAN          4
# }