Find current installations

Despite the description minimal requirements find which versions are required due to dependencies.

Usage

package_dependencies(pkg = ".", which = "strong")

Arguments

pkg

Path to a file with

which

a character vector listing the types of dependencies, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"). Character string "all" is shorthand for that vector, character string "most" for the same vector without "Enhances", character string "strong" (default) for the first three elements of that vector.

Value

A data.frame with the name, version required and version used.

Examples

pd <- package_dependencies("ggeasy")
#> Warning: Some dependencies are not on available repositories: cli, glue, gtable, isoband, lifecycle, MASS, mgcv, rlang, scales, tibble, vctrs, withr, nlme, Matrix, farver, labeling, munsell, R6, RColorBrewer, viridisLite, fansi, magrittr, pillar, pkgconfig, lattice, colorspace, utf8 .
#> Check for Additional_repositories or other repositories (Bioconductor.org?).
head(pd)
#>      name   op version    type package required
#> 1       R   >=   3.5.0 Depends  ggeasy     <NA>
#> 2 ggplot2 <NA>    <NA> Imports  ggeasy     <NA>
#> 3   rlang <NA>    <NA> Imports  ggeasy     <NA>