Resolve links

Converts Anchors and targets so that it can be easily understood. See WRE for extensive explanations

Usage

split_anchor(links, count = TRUE)

Arguments

links

A data.frame with Package, Source, Anchor and Target.

count

A logical value if links should be counted.

Value

A data.frame with Package, Source, to_pkg, to_target, n (number of times it happens)

Details

There are 4 different types of links:

{Target}

[=Target]{name}

[package]{Target}

[package:target]{name} The first two can be to any package and led to disambiguation pages, the last two are fully resolved (package and alias)

See also