Check versions

Helper function to detect which package have a required version on the dependencies that could be upgraded.

Usage

update_dependencies(deps)

Arguments

deps

The output of package_dependencies() or any data.frame with a version and required column.

Value

The data.frame filtered with some relevant rows

See also

Examples

pd <- package_dependencies("arrow")
#> Warning: Some dependencies are not on available repositories. Check for 'Additional_repositories' or other repositories (Bioconductor.org?):
#> aroma.light, DNAcopy, phyloseq, preprocessCore, pcaMethods, sparseMatrixStats, qvalue, Rgraphviz, edgeR, ALDEx2, EBImage, Biostrings, GenomicAlignments, IRanges, limma, org.Hs.eg.db, AnnotationDbi, DESeq2, Biobase, pwalign, maftools, ComplexHeatmap, GSVA, SummarizedExperiment, siggenes, annotate, fgsea, ConsensusClusterPlus, clusterProfiler, sva, affyPLM, affy, graph, fabia, RBGL, Rsamtools, impute, GO.db, biomaRt, rhdf5, mixOmics, survcomp, ggtree, treeio, GOSemSim, GENIE3, GenomicRanges, org.Dr.eg.db, org.Mm.eg.db, bluster, genefilter, BiocParallel, multtest, fmcsR, ChemmineR, ChIPseeker, TxDb.Hsapiens.UCSC.hg38.knownGene, TxDb.Hsapiens.UCSC.hg19.knownGene, TxDb.Mmusculus.UCSC.mm10.knownGene, SingleCellExperiment, TCseq, GSEABase, singscore, GenomicInteractions, rtracklayer, BiocFileCache, DECIPHER, tximport, BiocGenerics, GenomeInfoDb, S4Vectors, ensembldb, STRINGdb, GenomicFeatures, TCGAbiolinks, SNPRelate, snpStats, dada2, ShortRead, RedeR, Icens, RCy3, signatureSearch, signatureSearchData, MLInterfaces, ALL, TSCAN, MAST, DelayedArray, HDF5Array, minfi, IlluminaHumanMethylationEPICanno.ilm10b4.hg19, IlluminaHumanMethylation450kanno.ilmn12.hg19, AUCell, Rdisop, gdsfmt, BiocStyle, BSgenome, BSgenome.Hsapiens.UCSC.hg38, GEOquery, biocViews, reactome.db, ggmsa, VariantAnnotation, DOSE, Heatplus, fmrs, bamsignals, KEGGgraph, graphite, SparseArray, zlibbioc, Rhtslib, seqLogo, vsn, M3C, pathview, AnnotationFilter, QSutils, MassSpecWavelet, qusage, microbiome, InteractiveComplexHeatmap, globaltest, GlobalAncova, scran, scuttle, a4Core, ggkegg, ReactomePA, Mfuzz, BiocVersion, EmpiricalBrownsMethod, minet, scater, glmSparseNet, tweeDEseq, iBBiG, BicARE, R4RNA, AnnotationHub, beadarray, GSVAdata, rWikiPathways, glmGamPoi, UCell, batchelor, MatrixGenerics, ReactomeGSA, celldex, SingleR, splatter, widgetTools, Rhdf5lib, GWASTools, SpatialExperiment, zinbwave, DelayedMatrixStats, qpgraph, csaw, InteractionSet, HiCDOC, coRdon, KEGGREST
update_dependencies(pd)